+++ src/lib/components/Timer.svelte
@@ -1,25 +1,21 @@
 <script lang="ts">
     import { onDestroy, onMount } from "svelte";
+    import { formatTime } from "$lib/types";
     
     let { 
         time = $bindable(),
         timerOpen = $bindable(),
-        addGrid
+        duration = $bindable(),
+        onPomodoroEnd
     } : {
         time: number,
         timerOpen: boolean,
-        addGrid: () => void
+        duration: number,
+        onPomodoroEnd: () => void
     } = $props();
 
     let localTime = $state(time);
     let displayTime = $state("00:00:00");
-    
-    function formatTime(totalSeconds: number): string {
-        const minutes = Math.floor((totalSeconds % 3600) / 60);
-        const seconds = totalSeconds % 60;
-        
-        return `${String(minutes).padStart(2, '0')}:${String(seconds).padStart(2, '0')}`;
-    }
     
     onMount(() => {
         const timer = setInterval(() => {
@@ -37,16 +33,17 @@
     });
 
     onDestroy(() => {
-        addGrid()
+        duration = time-localTime
+        onPomodoroEnd()
     })
 </script>
 
 <div class="w-full flex flex-col items-center">
     <div class="w-full mt-10 flex flex-col items-center z-50">
-        <h1>{displayTime}</h1>
+        <h1 class="text-white" style="text-shadow: 2px 2px 4px rgba(0,0,0,0.5);">{displayTime}</h1>
 
         <!-- Bar -->
-        <ul id="menuBar" class="bg-pink-100 rounded-full p-2 flex flex-row">
+        <ul id="menuBar" class="bg-rose-100 rounded-full p-2 flex flex-row pointer-events-auto shadow-lg">
             <li onclick={() => timerOpen = false}>Stop</li>
         </ul>
     </div>